+++ app/security/rules/SecurityRule.scala
@@ -2,5 +2,5 @@ package security.rules
 
 trait SecurityRule
 case class ProjectOwner(hash: String) extends SecurityRule
-case class NodeOwner(nodeId: Int) extends SecurityRule
-case class NotChildNode(nodeId: Int, targetId: Int) extends SecurityRule
+case class NodeOwner(nodeId: Long) extends SecurityRule
+case class NotChildNode(nodeId: Long, targetId: Long) extends SecurityRule
