+++ app/v1/project/ProjectService.scala
@@ -1,13 +1,9 @@
 package v1.project
 
 import javax.inject._
-import com.mohiva.play.silhouette.impl.exceptions.AccessDeniedException
-import com.mohiva.play.silhouette.impl.providers.OAuth1Provider.AuthorizationError
-import utils.Implicits._
+import v1.generic.extensions.decodeHash
 import v1.project_user.ProjectUserRepo
 import v1.user.{User, UserRepo}
-import scalaz.Scalaz._
-import v1.generic.extensions.decodeHash
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
@@ -36,11 +32,11 @@ class ProjectService @Inject()(
         .fallbackTo(createProject(user))
     }
 
-    def userInProject(user: Int, project: Int): Future[Boolean] = {
+    def userInProject(user: Long, project: Long): Future[Boolean] = {
         puRepo.find(user, project).map(_.isDefined)
     }
 
-    def findByHashAndUser(hash: String, user: Int): Future[Project] = {
+    def findByHashAndUser(hash: String, user: Long): Future[Project] = {
         val Some(id) = decodeHash(hash)
         for {
             ok            <- userInProject(user, id)
