+++ app/v1/project_user/ProjectUser.scala
@@ -1,5 +1,5 @@
 package v1.project_user
 
-case class ProjectUser(user: Int, project: Int)
+case class ProjectUser(user: Long, project: Long)
