+++ app/v1/version/Version.scala
@@ -2,6 +2,6 @@ package v1.version
 
 import v1.generic.Entity
 
-case class Version(id: Int, name: String) extends Entity
+case class Version(id: Long, name: String) extends Entity
